refactor(definitions): refactor createEventWithPrompt (#166)
Rename interface to event operations.
